api/requirements.txt
beautifulsoup4==4.8.0
dropbox==9.3.0
Flask==1.0.2
Flask-Cors==3.0.7
flask-talisman==0.7.0
gunicorn==19.8.1
mongoengine==0.18.2
pymongo[tls,srv]==3.8.0
python-dateutil==2.8.0
python-dotenv==0.13.0
requests==2.22.0
